move inventory.py to forseti
